// src/interfaces/FavoritesResponseData.ts
export interface FavoritesResponseData {
  items: [
    {
      id: "140a6a2e-edcc-40b5-b43e-2f64432671bf";
      zoomInfoId: any;
      name: "Kinross";
      logo: any;
      ticker: "NYSE: KGC";
      parentCompany: any;
      phone: "(416) 365-5123";
      fax: "(416) 363-6622";
      website: "www.kinross.com";
      city: "Toronto";
      street: "25 York St, 17th Fl";
      state: "Ontario";
      zipCode: "M5J 2V5";
      country: "Canada";
      continent: any;
      productsBrandDescription: any;
      descriptionList: "Founded in 1993, Kinross is a Canadian-based senior gold mining company with mines and projects in the United States, Brazil, Russia, Mauritania, Chile, and Ghana. Their focus is on delivering value based on the core principles of operational excellence, balance sheet strength, disciplined growth and responsible mining. Kinross maintains listings on the Toronto Stock Exchange (TSX: K) and the New York Stock Exchange (NYSE: KGC).";
      revenueRange: "$1 bil. - $5 bil.";
      employeeRange: "35000";
      twitterHandle: any;
      socialMediaUrls: any;
      competitors: any;
      subUnitIndustries: any;
      primaryIndustry: [any];
      industries: any;
      revenue: "3500000";
      employeeCount: 175000;
      annualContributions: any;
      cashContributions: any;
      inKindContributions: any;
      employeeContributions: any;
      parentId: any;
      parentName: any;
      type: any;
      sdgGoals: [];
      genders: any;
      income: any;
      age: any;
      ethnicity: any;
      nonprofit: any;
      purchase: any;
      affiliation: any;
      brands: any;
      interests: any;
      typesOfInvestment: any;
      errorLoadZoomInfo: any;
      charitablePartners: [];
      statusZoomInfo: "no_loaded";
      loadZoomInfoDate: any;
      errorLoadZoomInfoDate: any;
      partnershipLink: any;
      employeeEngagementOpportunities: true;
      similarCompanies: [
        "IAMGOLD, Agnico Eagle Mines, Endeavour Silver, Yamana Gold, Hecla Mining, Barrick Gold, Dundee Precious Metals, New Gold, Centerra Gold, Newmont Corporation, Fortuna Silver Mines, Polymetal International, Alamos Gold, Gold Fields, Teranga Gold, Kennecott Utah Copper, Banro, Eldorado Gold, Fresnillo, Kirkland Lake Gold, B2Gold, KGHM Polska Mied?, SSR Mining, HudBay Minerals, Argonaut Gold"
      ];
      favoriteCompanies: [
        {
          id: "34369898-eb9f-407f-ac45-8e28ada8f7b1";
          companyId: "140a6a2e-edcc-40b5-b43e-2f64432671bf";
          userId: "eb893b0e-ce28-4392-a824-4a3ff9c36e79";
        }
      ];
      score: 50;
      like: true;
      crsFocus: [];
    }
  ];
  meta: {
    totalItems: 1;
    itemCount: 1;
    itemsPerPage: "1";
    totalPages: 1;
    currentPage: "1";
  };
}